Cut out the function of sampling into another script
